proposal: Move elektraKsToMemArray to kdbease
