+++ tests/ctest/CMakeLists.txt
@@ -23,6 +23,7 @@ target_link_elektra (test_splitset elektra-plugin)
 
 target_link_elektra (test_meta elektra-meta)
 target_link_elektra (test_meta elektra-proposal)
+target_link_elektra (test_ks elektra-ease)
 target_link_elektra (test_key elektra-meta)
 target_link_elektra (test_key elektra-proposal)
